Refactored ships
